Related work items: #339

+++ src/modules/Assessment/containers/AssessmentList.js
@@ -5,8 +5,14 @@ import { Button } from "reactstrap";
 import { ContentLayout, ModalLayout, TableLayout } from "shared";
 import { LoadingSpinner } from "shared/components/LoadingSpinner";
 import { useAssessment } from "../hooks";
+import { useAuthorizeContext } from "master";
 
-export const AssessmentList = (props) => {
+export const AssessmentList = () => {
+  const { hasPermission } = useAuthorizeContext();
+  const hasCreatePermission = hasPermission("Assessments", "Create");
+  const hasEditPermission = hasPermission("Assessments", "Edit");
+  const hasDeletePermission = hasPermission("Assessments", "Delete");
+
   const {
     assessmentQuery,
     isConfirmDelete,
@@ -17,6 +23,7 @@ export const AssessmentList = (props) => {
   } = useAssessment({
     load: true,
   });
+
   const { data, isLoading } = assessmentQuery;
   const history = useNavigate();
   const location = useLocation();
@@ -36,12 +43,16 @@ export const AssessmentList = (props) => {
   const ActionButtons = ({ row }) => {
     return (
       <>
-        <Button outline color="primary" size="sm" onClick={() => onEdit(row.original.assessmentId)}>
-          <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
-        </Button>
-        <Button color="danger" size="sm" onClick={() => onDelete(row)} className="ms-3">
-          <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
-        </Button>
+        {
+          hasEditPermission && <Button outline color="primary" size="sm" onClick={() => onEdit(row.original.assessmentId)}>
+            <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
+          </Button>
+        }
+        {
+          hasDeletePermission && <Button color="danger" size="sm" onClick={() => onDelete(row)} className="ms-3">
+            <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
+          </Button>
+        }
       </>
     );
   };
@@ -73,13 +84,16 @@ export const AssessmentList = (props) => {
       Header: "Pass Mark",
       accessor: "passMark",
     },
-    {
+  ];
+
+  if (hasEditPermission || hasDeletePermission) {
+    columns.push({
       Header: "Actions",
       id: "actions",
       accessor: "assessmentId",
       Cell: ActionButtons,
-    },
-  ];
+    })
+  }
 
   if (isLoading) {
     return <LoadingSpinner />;
@@ -88,11 +102,13 @@ export const AssessmentList = (props) => {
   return (
     <>
       <ContentLayout title="Assessments" subtitle="List" breadcrumb={[{ label: "Assessments" }]}>
-        <div className="mb-4">
-          <Button color="primary" size="sm" onClick={gotoCreate}>
-            Create New
-          </Button>
-        </div>
+        {
+          hasCreatePermission && <div className="mb-4">
+            <Button color="primary" size="sm" onClick={gotoCreate}>
+              Create New
+            </Button>
+          </div>
+        }
         <TableLayout columns={columns} data={data} />
         <ModalLayout
           isOpen={isConfirmDelete}
