+++ src/modules/Batch/containers/BatchList.js
@@ -6,8 +6,15 @@ import { Button } from "reactstrap";
 import { useNavigate, useLocation } from "react-router-dom";
 import { useBatch } from "../hooks";
 import { LoadingSpinner } from "shared/components/LoadingSpinner";
+import { useAuthorizeContext } from "master";
+
 
 export const BatchList = () => {
+  const { hasPermission } = useAuthorizeContext();
+  const hasCreatePermission = hasPermission("Batches", "Create");
+  const hasEditPermission = hasPermission("Batches", "Edit");
+  const hasDeletePermission = hasPermission("Batches", "Delete");
+
   const { batch, batchesQuery, onDelete, onToggleModal, isConfirmDelete, deleteBatch } = useBatch({
     load: true,
   });
@@ -26,17 +33,21 @@ export const BatchList = () => {
   const ActionButtons = ({ value }) => {
     return (
       <>
-        <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
-          <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
-        </Button>
-        <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
-          <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
-        </Button>
+        {
+          hasEditPermission && <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
+            <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
+          </Button>
+        }
+        {
+          hasDeletePermission && <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
+            <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
+          </Button>
+        }
       </>
     );
   };
 
-  const columns = [
+  let columns = [
     {
       Header: "Name",
       accessor: "name",
@@ -49,14 +60,18 @@ export const BatchList = () => {
       Header: "End",
       accessor: "endDate",
     },
-
-    {
-      Header: "Actions",
-      accessor: "batchId",
-      id: "actions",
-      Cell: ActionButtons,
-    },
   ];
+
+  if (hasEditPermission || hasDeletePermission) {
+    columns.push(
+      {
+        Header: "Actions",
+        accessor: "batchId",
+        id: "actions",
+        Cell: ActionButtons,
+      }
+    );
+  }
 
   if (isLoading) {
     return <LoadingSpinner />;
@@ -68,7 +83,9 @@ export const BatchList = () => {
       subtitle={"List"}
       breadcrumb={[{ label: "Batch", location: "/batch" }]}
     >
-      <BatchFilter />
+      {
+        hasCreatePermission && <BatchFilter />
+      }
       <TableLayout columns={columns} data={data} />
       <ModalLayout
         isOpen={isConfirmDelete}
