+++ src/modules/Courses/containers/CourseList.js
@@ -5,14 +5,20 @@ import { Button } from "reactstrap";
 import { useNavigate, useLocation } from "react-router-dom";
 import { useCourse } from "../hooks";
 import { LoadingSpinner } from "shared/components/LoadingSpinner";
+import { useAuthorizeContext } from "master";
 
-export const CourseList = (props) => {
+export const CourseList = () => {
+  const history = useNavigate();
+  const location = useLocation();
+  const { hasPermission } = useAuthorizeContext();
+  const hasDeletePermission = hasPermission("Courses", "Delete");
+  const hasEditPermission = hasPermission("Courses", "Edit");
+  const hasCreatePermission = hasPermission("Courses", "Create");
   const { course, coursesQuery, onDelete, onToggleModal, isConfirmDelete, deleteCourse } =
     useCourse({
       load: true,
     });
-  const history = useNavigate();
-  const location = useLocation();
+
 
   const onConfirm = () => {
     deleteCourse.mutate(course.courseId);
@@ -37,52 +43,61 @@ export const CourseList = (props) => {
   const ActionButtons = ({ value }) => {
     return (
       <>
-        <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
-          <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
-        </Button>
-        <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
-          <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
-        </Button>
+        {
+          hasEditPermission && <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
+            <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
+          </Button>
+        }
+        {
+          hasDeletePermission && <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
+            <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
+          </Button>
+        }
       </>
     );
   };
 
-  const columns = useMemo(
-    () => [
-      {
-        Header: "Name",
-        accessor: "name",
-      },
-      {
-        Header: "Exam",
-        accessor: "hasExam",
-        Cell: CheckMarker,
-      },
-      {
-        Header: "Content",
-        accessor: "isContentEnabled",
-        Cell: CheckMarker,
-      },
-      {
-        Header: "Thumbnail",
-        accessor: "courseImageURL",
-        Cell: Thumbnail,
-      },
-      {
-        Header: "Actions",
-        id: "actions",
-        accessor: "courseId",
-        Cell: ActionButtons,
-      },
-    ],
-    []
-  );
+  const columns = [
+    {
+      Header: "Name",
+      accessor: "name",
+    },
+    {
+      Header: "Exam",
+      accessor: "hasExam",
+      Cell: CheckMarker,
+    },
+    {
+      Header: "Content",
+      accessor: "isContentEnabled",
+      Cell: CheckMarker,
+    },
+    {
+      Header: "Thumbnail",
+      accessor: "courseImageURL",
+      Cell: Thumbnail,
+    },
+
+  ];
+
+  if (hasEditPermission || hasDeletePermission) {
+    columns.push({
+      Header: "Actions",
+      id: "actions",
+      accessor: "courseId",
+      Cell: ActionButtons,
+    });
+  }
+
   if (coursesQuery.isLoading) {
     return <LoadingSpinner />;
   }
+
   return (
     <ContentLayout title={"Courses"} subtitle={"List"} breadcrumb={[{ label: "Courses" }]}>
-      <CourseIdFilter />
+      {
+        hasCreatePermission && <CourseIdFilter />
+      }
       <TableLayout columns={columns} data={coursesQuery.data} />
       <ModalLayout
         isOpen={isConfirmDelete}
