+++ src/modules/SupportTickets/containers/SupportTicketList.js
@@ -1,13 +1,20 @@
-import React, { useMemo } from "react";
-import { useTable } from "react-table";
+import React from "react";
 import { ContentLayout, TableLayout, ModalLayout } from "shared/components";
 import { SupportTicketFilter } from "..";
 import { Button, Badge } from "reactstrap";
 import { useNavigate, useLocation } from "react-router-dom";
 import { useSupportTicket } from "../hooks";
 import { LoadingSpinner } from "shared/components/LoadingSpinner";
+import { useAuthorizeContext } from "master";
+
+
 
 export const SupportTicketList = () => {
+  const { hasPermission } = useAuthorizeContext();
+  const hasCreatePermission = hasPermission("SupportTickets", "Create");
+  const hasEditPermission = hasPermission("SupportTickets", "Edit");
+  const hasDeletePermission = hasPermission("SupportTickets", "Delete");
+
   const {
     supportTicket,
     supportTicketsQuery,
@@ -41,12 +48,16 @@ export const SupportTicketList = () => {
   const ActionButtons = ({ value }) => {
     return (
       <>
-        <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
-          <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
-        </Button>
-        <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
-          <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
-        </Button>
+        {
+          hasEditPermission && <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
+            <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
+          </Button>
+        }
+        {
+          hasDeletePermission && <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
+            <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
+          </Button>
+        }
       </>
     );
   };
@@ -69,13 +80,17 @@ export const SupportTicketList = () => {
       accessor: "status",
       Cell: StatusIndicator,
     },
-    {
+
+  ];
+
+  if (hasDeletePermission || hasEditPermission) {
+    columns.push({
       Header: "Actions",
       accessor: "ticketId",
       id: "actions",
       Cell: ActionButtons,
-    },
-  ];
+    });
+  }
 
   if (isLoading) {
     return <LoadingSpinner />;
@@ -87,7 +102,9 @@ export const SupportTicketList = () => {
       subtitle={"List"}
       breadcrumb={[{ label: "Support ticket", location: "/supportTicket" }]}
     >
-      <SupportTicketFilter />
+      {
+        hasCreatePermission && <SupportTicketFilter />
+      }
       <TableLayout columns={columns} data={data} />
       <ModalLayout
         isOpen={isConfirmDelete}
