+++ src/modules/TrainingPartner/containers/TrainingPartnerList.js
@@ -4,8 +4,14 @@ import { useTrainingPartner } from "../hooks";
 import { useNavigate, useLocation } from "react-router-dom";
 import { Button } from "reactstrap";
 import { LoadingSpinner } from "shared/components/LoadingSpinner";
+import { useAuthorizeContext } from "master";
 
 export const TrainingPartnerList = () => {
+  const { hasPermission } = useAuthorizeContext();
+  const hasCreatePermission = hasPermission("TrainingPartner", "Create");
+  const hasEditPermission = hasPermission("TrainingPartner", "Edit");
+  const hasDeletePermission = hasPermission("TrainingPartner", "Delete");
+
   const {
     trainingPartner,
     trainingPartnerQuery,
@@ -33,12 +39,16 @@ export const TrainingPartnerList = () => {
   const ActionButtons = ({ value }) => {
     return (
       <>
-        <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
-          <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
-        </Button>
-        <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
-          <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
-        </Button>
+        {
+          hasEditPermission && <Button outline color="primary" size="sm" onClick={() => onEdit(value)}>
+            <i className="bi bi-pencil-square" style={{ fontSize: "10px" }}></i> <span>Edit</span>
+          </Button>
+        }
+        {
+          hasDeletePermission && <Button color="danger" size="sm" onClick={() => onDelete(value)} className="ms-3">
+            <i className="bi bi-trash" style={{ fontSize: "10px" }}></i> <span>Delete</span>
+          </Button>
+        }
       </>
     );
   };
@@ -56,14 +66,17 @@ export const TrainingPartnerList = () => {
       Header: "E-Mail",
       accessor: "emailAddress",
     },
-
-    {
-      Header: "Actions",
-      accessor: "trainingPartnerId",
-      id: "actions",
-      Cell: ActionButtons,
-    },
   ];
+
+  if (hasDeletePermission || hasEditPermission) {
+    columns.push(
+      {
+        Header: "Actions",
+        accessor: "trainingPartnerId",
+        id: "actions",
+        Cell: ActionButtons,
+      });
+  }
 
   if (isLoading) {
     return <LoadingSpinner />;
@@ -75,10 +88,10 @@ export const TrainingPartnerList = () => {
       subtitle={"Training Partner List"}
       breadcrumb={[{ label: "Training Partner", location: "/trainingpartner" }]}
     >
-      <TrainingPartnerFilter />
-
+      {
+        hasCreatePermission && <TrainingPartnerFilter />
+      }
       <TableLayout columns={columns} data={data} />
-
       <ModalLayout
         isOpen={isConfirmDelete}
         title={"Confirm"}
